+++ MP1/src/astar_average.hpp
@@ -28,7 +28,7 @@ namespace astar {
         std::string name() const;
         
         // distance measure
-        unsigned int operator()( const multi::state & s ) const;
+        double operator()( const multi::state & s ) const;
         
         void setScaleFactor( double scale );
