+++ MP1/src/astar_euclidean.cpp
@@ -22,11 +22,11 @@ namespace astar {
         return "Manhattan Distance";
     }
 
-    unsigned int euclidean_dist::operator()( const multi::state & s ) const {
+    double euclidean_dist::operator()( const multi::state & s ) const {
         
         // setup variables being used
         maze::id_type node1 = s.current_node;
-        unsigned int cost   = UINT32_MAX;
+        double cost   = UINT32_MAX;
         const maze* maze_   = this->getMaze();
         const gplist* gplist_= this->getGoalPointList();
