+++ MP1/src/astar_manhattan.cpp
@@ -25,11 +25,11 @@ namespace astar {
         return "Manhattan Distance";
     }
     
-    unsigned int manhattan_dist::operator()( const multi::state & s ) const {
+    double manhattan_dist::operator()( const multi::state & s ) const {
         
         // setup variables being used
         maze::id_type node1 = s.current_node;
-        unsigned int cost   = UINT32_MAX;
+        double cost   = UINT32_MAX;
         const maze* maze_   = this->getMaze();
         const gplist* gplist_= this->getGoalPointList();
